Refactor MoreShowsPresenterFactory to use Kotlin Inject:
- Extract interface and default implementation; update imports and usages.
- Replace test implementation with FakeMoreShowsPresenterFactory in DefaultRootComponentTest.
