+++ presenter/more-shows/src/commonMain/kotlin/com/thomaskioko/tvmaniac/moreshows/presentation/MoreShowsPresenter.kt
@@ -28,30 +28,6 @@ import me.tatarka.inject.annotations.Assisted
 import me.tatarka.inject.annotations.Inject
 
 @Inject
-class MoreShowsPresenterFactory(
-    private val popularShowsRepository: PopularShowsRepository,
-    private val upcomingShowsRepository: UpcomingShowsRepository,
-    private val trendingShowsRepository: TrendingShowsRepository,
-    private val topRatedShowsRepository: TopRatedShowsRepository,
-) {
-    fun create(
-        componentContext: ComponentContext,
-        id: Long,
-        onBack: () -> Unit,
-        onNavigateToShowDetails: (id: Long) -> Unit,
-    ): MoreShowsPresenter = MoreShowsPresenter(
-        componentContext = componentContext,
-        categoryId = id,
-        onBack = onBack,
-        onNavigateToShowDetails = onNavigateToShowDetails,
-        popularShowsRepository = popularShowsRepository,
-        upcomingShowsRepository = upcomingShowsRepository,
-        trendingShowsRepository = trendingShowsRepository,
-        topRatedShowsRepository = topRatedShowsRepository,
-    )
-}
-
-@Inject
 class MoreShowsPresenter(
     @Assisted componentContext: ComponentContext,
     @Assisted private val categoryId: Long,
